# Conflicts:
#	README.md
#	be/src/.env
#	be/src/Infrastructure/Localization/Foundation.Localization/Resources/GlobalResources.de.resx
#	be/src/Infrastructure/Localization/Foundation.Localization/Resources/GlobalResources.en.resx
#	be/src/Infrastructure/Localization/Foundation.Localization/Resources/GlobalResources.pl.resx
#	be/src/Infrastructure/Mailing/Foundation.Mailing/Configurations/MailingConfiguration.cs
#	be/src/Project/Services/Identity/Identity.Api/wwwroot/dist/js/resetpasswordpage.js
#	be/src/Project/Services/Identity/Identity.Api/wwwroot/dist/js/setpasswordpage.js
#	be/src/Project/Services/Identity/Identity.Api/wwwroot/dist/js/signinpage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/css/statusOrderPage.css
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/availableproductspage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/categorypage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/homepage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/listOrdersPage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/newOrderPage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/newsItemPage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/newsPage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/outletpage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/productpage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/searchproductspage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/statusOrderPage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/css/editorderpage.css
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/categoriespage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/categorypage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/clientgrouppage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/clientgroupspage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/clientpage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/clientspage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/editorderpage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/inventoriespage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/inventorypage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/newscategoriespage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/newscategorypage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/newsitempage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/newspage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/orderpage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/orderspage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/outletpage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/outletspage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/productattributeitempage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/productattributepage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/productattributespage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/productpage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/productspage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/settingspage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/warehousepage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/warehousespage.js
#	fe/src/project/Seller.Portal/areas/Orders/components/EditOrderForm/EditOrderForm.js

+++ be/src/Project/Services/Inventory/Inventory.Api/ServicesModels/InventoryServiceModels/InventorySumServiceModel.cs
@@ -9,8 +9,8 @@ namespace Inventory.Api.ServicesModels.InventoryServiceModels
         public string ProductName { get; set; }
         public string ProductSku { get; set; }
         public string ProductEan { get; set; }
-        public int? Quantity { get; set; }
-        public int? AvailableQuantity { get; set; }
+        public double? Quantity { get; set; }
+        public double? AvailableQuantity { get; set; }
         public int? RestockableInDays { get; set; }
         public IEnumerable<InventoryServiceModel> Details { get; set; }
         public DateTime? ExpectedDelivery { get; set; }
