+++ be/src/Project/Services/Inventory/Inventory.Api/ServicesModels/InventoryServiceModels/UpdateProductInventoryServiceModel.cs
@@ -10,8 +10,8 @@ namespace Inventory.Api.ServicesModels.InventoryServiceModels
         public string ProductName { get; set; }
         public string ProductSku { get; set; }
         public string ProductEan { get; set; }
-        public int Quantity { get; set; }
-        public int AvailableQuantity { get; set; }
+        public double Quantity { get; set; }
+        public double AvailableQuantity { get; set; }
         public int? RestockableInDays { get; set; }
         public DateTime? ExpectedDelivery { get; set; }
     }
