+++ be/src/Project/Services/Inventory/Inventory.Api/ServicesModels/OutletServiceModels/UpdateBasketOutletServiceModel.cs
@@ -6,6 +6,6 @@ namespace Inventory.Api.ServicesModels.OutletServiceModels
     public class UpdateBasketOutletServiceModel : BaseServiceModel
     {
         public Guid? ProductId { get; set; }
-        public int BookedQuantity { get; set; }
+        public double BookedQuantity { get; set; }
     }
 }
