+++ be/src/Project/Services/Inventory/Inventory.Api/v1/ResponseModels/OutletResponseModel.cs
@@ -10,11 +10,11 @@ namespace Inventory.Api.v1.ResponseModels
         public string ProductSku { get; set; }
         public Guid? WarehouseId { get; set; }
         public string WarehouseName { get; set;}
-        public int Quantity { get; set; }
+        public double Quantity { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
         public string Ean { get; set; }
-        public int? AvailableQuantity { get; set; }
+        public double? AvailableQuantity { get; set; }
         public DateTime LastModifiedDate { get; set; }
         public DateTime CreatedDate { get; set; }
