+++ be/src/Project/Web/Seller/Seller.Web/Areas/Inventory/Repositories/Inventories/InventoryRepository.cs
@@ -214,7 +214,7 @@ namespace Seller.Web.Areas.Inventory.Repositories.Inventories
             }
         }
 
-        public async Task<Guid> SaveAsync(string token, string language, Guid? id, Guid? WarehouseId, Guid? ProductId, string ProductName, string ProductSku, int Quantity, string ean, int? RestockableInDays, int? AvailableQuantity, DateTime? ExpectedDelivery, Guid? OrganisationId)
+        public async Task<Guid> SaveAsync(string token, string language, Guid? id, Guid? WarehouseId, Guid? ProductId, string ProductName, string ProductSku, double Quantity, string ean, int? RestockableInDays, double? AvailableQuantity, DateTime? ExpectedDelivery, Guid? OrganisationId)
         {
             var requestModel = new SaveInventoryRequestModel
             {
